test: Refine FileHandle open tests & fix Windows append-at-end semantics
* Add tests covering more scenarios for opening File Handle
* Add tests for correct append behaviour with no-follow + truncate on Windows
* Fix the append method of `AppendHandle` on Windows, which may lose atomic append behaviour when truncate=true

+++ Sources/SwiftFileSystem/FileHandle/ReadWriteFileHandle.swift
@@ -155,7 +155,7 @@ extension ReadWriteFileHandle {
             }
         }
 
-        if options.noFollow && options.truncate && type == .regular {
+        if options.noFollow && options.truncate && creationOption != .assertMissing && type == .regular {
             try catchLowLevelError(operation: .open(path)) { () throws(LowLevelError) in
                 try handle.truncate()
             }
